+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Register `Replication` type and add `#[reflect(Component)]`.
+
 ## [0.15.0] - 2023-10-21
 
 ### Added